+++ src/main/java/io/github/eggy03/ferrumx/windows/entity/storage/Win32DiskDrive.java
@@ -75,7 +75,7 @@ public class Win32DiskDrive {
 
     @SerializedName("Partitions")
     @Nullable
-    Integer partitions;
+    Long partitions;
 
     @SerializedName("Status")
     @Nullable
